+++ driver/web/adapter.go
@@ -49,7 +49,7 @@ type Adapter struct {
 
 // @title Rokwire Health Building Block API
 // @description Rokwire Health Building Block API Documentation.
-// @version 1.13.0
+// @version 1.14.0
 // @license.name Apache 2.0
 // @license.url http://www.apache.org/licenses/LICENSE-2.0.html
 // @host localhost
@@ -487,8 +487,8 @@ func (we Adapter) providerAuthWrapFunc(handler http.HandlerFunc) http.HandlerFun
 
 //NewWebAdapter creates new WebAdapter instance
 func NewWebAdapter(host string, app *core.Application, appKeys []string, oidcProvider string,
-	oidcAppClientID string, oidcAdminClientID string, phoneAuthSecret string, providersKeys []string) Adapter {
-	auth := NewAuth(app, appKeys, oidcProvider, oidcAppClientID, oidcAdminClientID, phoneAuthSecret, providersKeys)
+	oidcAppClientID string, adminAppClientID string, adminWebAppClientID string, phoneAuthSecret string, providersKeys []string) Adapter {
+	auth := NewAuth(app, appKeys, oidcProvider, oidcAppClientID, adminAppClientID, adminWebAppClientID, phoneAuthSecret, providersKeys)
 	authorization := casbin.NewEnforcer("driver/web/authorization_model.conf", "driver/web/authorization_policy.csv")
 
 	apisHandler := rest.NewApisHandler(app)